+++ projects/framework-common-extensions/release_notes.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 
+* [changed] Add windows specific titles on standard opener dialog and standard publisher dialog.
 * [fixed] Disabled auto scrolling on asset list. Always select the first item on activation.
 * [changed] Added labels to publisher asset selector and publisher, aligned margins.
 * [new] Initial release.